test: group DCP resume coverage
